requirements.txt
Django==1.7.1
Pygments==2.0.1
argparse==1.2.1
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-cors-headers==1.0.0
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==3.0.2
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
